+++ vector/src/main/java/im/vector/app/core/epoxy/bottomsheet/BottomSheetMessagePreviewItem.kt
@@ -19,6 +19,7 @@ package im.vector.app.core.epoxy.bottomsheet
 import android.text.method.MovementMethod
 import android.widget.ImageView
 import android.widget.TextView
+import androidx.core.view.isVisible
 import com.airbnb.epoxy.EpoxyAttribute
 import com.airbnb.epoxy.EpoxyModelClass
 import im.vector.app.R
@@ -27,6 +28,7 @@ import im.vector.app.core.epoxy.VectorEpoxyModel
 import im.vector.app.core.extensions.setTextOrHide
 import im.vector.app.features.home.AvatarRenderer
 import im.vector.app.features.home.room.detail.timeline.tools.findPillsAndProcess
+import im.vector.app.features.media.ImageContentRenderer
 import org.matrix.android.sdk.api.util.MatrixItem
 
 /**
@@ -45,6 +47,12 @@ abstract class BottomSheetMessagePreviewItem : VectorEpoxyModel<BottomSheetMessa
     lateinit var body: CharSequence
 
     @EpoxyAttribute
+    var imageContentRenderer: ImageContentRenderer? = null
+
+    @EpoxyAttribute
+    var data: ImageContentRenderer.Data? = null
+
+    @EpoxyAttribute
     var time: CharSequence? = null
 
     @EpoxyAttribute
@@ -59,10 +67,19 @@ abstract class BottomSheetMessagePreviewItem : VectorEpoxyModel<BottomSheetMessa
         holder.avatar.setOnClickListener { userClicked?.invoke() }
         holder.sender.setOnClickListener { userClicked?.invoke() }
         holder.sender.setTextOrHide(matrixItem.displayName)
+        data?.let {
+            imageContentRenderer?.render(it, ImageContentRenderer.Mode.THUMBNAIL, holder.imagePreview)
+        }
+        holder.imagePreview.isVisible = data != null
         holder.body.movementMethod = movementMethod
         holder.body.text = body
         body.findPillsAndProcess(coroutineScope) { it.bind(holder.body) }
         holder.timestamp.setTextOrHide(time)
+    }
+
+    override fun unbind(holder: Holder) {
+        imageContentRenderer?.clear(holder.imagePreview)
+        super.unbind(holder)
     }
 
     class Holder : VectorEpoxyHolder() {
@@ -70,5 +87,6 @@ abstract class BottomSheetMessagePreviewItem : VectorEpoxyModel<BottomSheetMessa
         val sender by bind<TextView>(R.id.bottom_sheet_message_preview_sender)
         val body by bind<TextView>(R.id.bottom_sheet_message_preview_body)
         val timestamp by bind<TextView>(R.id.bottom_sheet_message_preview_timestamp)
+        val imagePreview by bind<ImageView>(R.id.bottom_sheet_message_preview_image)
     }
 }