+++ CHANGES.md
@@ -9,6 +9,7 @@ Improvements 🙌:
  - Crypto improvement | Bulk send NO_OLM withheld code
  - Display the room shield in all room setting screens
  - Improve message with Emoji only detection (#3017)
+ - Picture preview when replying. Also add the image preview in the message detail bottomsheet (#2916)
  - Api interceptor to allow app developers peek responses (#2986)
 
 Bugfix 🐛: